+++ packages/core/src/evaluation/index.ts
@@ -1,5 +1,6 @@
 export * from "./external.js";
 export * from "./optimization.js";
+export * from "./proof.js";
 export * from "./static.js";
 export * from "./suite.js";
 export * from "./verification.js";